requirements.txt
datasets==2.17.0
transformers==4.38.0
streamlit
streamlit_option_menu
torch
mysql.connector
pandas